requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.1
Events==0.5
grpcio==1.78.0
idna==3.11
numpy==2.4.4
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
pgvector==0.4.2
protobuf==7.34.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.33.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
